refactor(cfd-2d): Decompose serpentine_flow/mod.rs into leaf modules
Part of ATLAS-ARCH-007.
