+++ crates/cfd-2d/src/solvers/serpentine_flow/mod.rs
@@ -86,527 +86,12 @@
 //! monotonically. Convergence is guaranteed by the spectral radius of the iteration matrix
 //! being strictly less than 1.
 
+mod geometry;
+mod mixing;
 mod solver;
+mod validation;
 
+pub use geometry::SerpentineGeometry;
+pub use mixing::{AdvectionDiffusionMixing, SerpentineMixingSolution};
 pub use solver::SerpentineSolver2D;
-
-use crate::scalar::Cfd2dScalar;
-use crate::scalar::{self, from_f64};
-use eunomia::{FloatElement, NumericElement};
-use serde::{Deserialize, Serialize};
-use std::f64::consts::PI;
-
-// ============================================================================
-// Serpentine Channel Geometry
-// ============================================================================
-
-/// Serpentine channel geometry with periodic turns
-///
-/// Defines a channel that alternates between straight sections and 90° turns.
-/// The path creates a snake-like pattern that enhances mixing.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SerpentineGeometry<T: Cfd2dScalar + Copy> {
-    /// Channel width \[m]
-    pub width: T,
-    /// Channel height \[m] (constant)
-    pub height: T,
-    /// Straight section length \[m]
-    pub l_straight: T,
-    /// Turn radius \[m] (for curved turns)
-    pub turn_radius: T,
-    /// Number of complete back-and-forth cycles
-    pub n_cycles: usize,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> SerpentineGeometry<T> {
-    /// Create standard serpentine for microfluidics
-    ///
-    /// # Typical Design
-    ///
-    /// - Width: 100-500 μm
-    /// - Straight sections: 500 μm
-    /// - Turn radius: 200 μm
-    /// - Cycles: 5-10 (for ~5-10 mm total length)
-    pub fn microfluidic_standard() -> Self {
-        Self {
-            width: from_f64::<T>(200e-6),
-            height: from_f64::<T>(50e-6),
-            l_straight: from_f64::<T>(500e-6),
-            turn_radius: from_f64::<T>(200e-6),
-            n_cycles: 5,
-        }
-    }
-
-    /// Create custom serpentine
-    pub fn new(width: T, height: T, l_straight: T, turn_radius: T, n_cycles: usize) -> Self {
-        Self {
-            width,
-            height,
-            l_straight,
-            turn_radius,
-            n_cycles,
-        }
-    }
-
-    /// Calculate total channel length
-    ///
-    /// Length = n_cycles × (2 × l_straight + turn_length)
-    /// where turn_length ≈ π × turn_radius (90° arc)
-    pub fn total_length(&self) -> T {
-        let pi = from_f64::<T>(PI);
-        let two = from_f64::<T>(2.0);
-        let turn_length = pi / from_f64::<T>(2.0) * self.turn_radius;
-
-        scalar::from_usize::<T>(self.n_cycles) * (two * self.l_straight + turn_length)
-    }
-
-    /// Get cross-sectional area
-    pub fn cross_section_area(&self) -> T {
-        self.width * self.height
-    }
-
-    /// Calculate number of diffusion lengths in straight section
-    ///
-    /// # Definition
-    ///
-    /// ```text
-    /// n_diff = (l_straight / width) × Pe
-    /// ```
-    ///
-    /// where Pe = u·w/D (Peclet number)
-    pub fn diffusion_lengths_per_section(&self, peclet: T) -> T {
-        let three = from_f64::<T>(3.0);
-        (self.l_straight / (self.width + from_f64::<T>(1e-15))) * three * peclet
-    }
-
-    /// Get bounding box [min_x, max_x, min_y, max_y]
-    pub fn bounding_box(&self) -> [T; 4] {
-        let r = self.turn_radius;
-        let ls = self.l_straight;
-        let w = self.width;
-        let half_w = w / from_f64::<T>(2.0);
-        let four_r = r * from_f64::<T>(4.0);
-
-        [
-            -r - half_w,
-            ls + r + half_w,
-            scalar::zero(),
-            four_r * scalar::from_usize::<T>(self.n_cycles),
-        ]
-    }
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement + std::ops::Rem<Output = T>> SerpentineGeometry<T> {
-    /// Check if a point (x, y) is within the fluid domain
-    ///
-    /// Assuming serpentine starts at (0, R) and snakes upwards in Y.
-    /// Channel centerline path:
-    /// Cycle N:
-    ///   Segment 1: (0, 4NR + R) -> (Ls, 4NR + R)
-    ///   Turn 1: Semi-circle around (Ls, 4NR + 2R) with radius R
-    ///   Segment 2: (Ls, 4NR + 3R) -> (0, 4NR + 3R)
-    ///   Turn 2: Semi-circle around (0, 4NR + 4R) with radius R
-    pub fn contains(&self, x: T, y: T) -> bool {
-        let r = self.turn_radius;
-        let ls = self.l_straight;
-        let w = self.width;
-        let half_w = w / from_f64::<T>(2.0);
-        let four_r = r * from_f64::<T>(4.0);
-
-        // Normalize y to cycle 0
-        let y_in_cycle = y % four_r;
-
-        let inner_r = r - half_w;
-        let outer_r = r + half_w;
-        let inner_r_sq = inner_r * inner_r;
-        let outer_r_sq = outer_r * outer_r;
-
-        // Sequence within a cycle:
-        // 0..2R: Segment 1 (bottom straight) and Turn 1 (right)
-        // 2R..4R: Segment 2 (top straight) and Turn 2 (left)
-
-        if y_in_cycle < r + half_w && y_in_cycle > r - half_w {
-            // Segment 1: (0, R) -> (Ls, R)
-            // Extend to x = -R for inlet tail
-            if x >= -r - half_w && x <= ls {
-                return true;
-            }
-        }
-
-        if y_in_cycle < from_f64::<T>(3.0) * r + half_w
-            && y_in_cycle > from_f64::<T>(3.0) * r - half_w
-        {
-            // Segment 2: (Ls, 3R) -> (0, 3R)
-            if x >= scalar::zero() && x <= ls {
-                return true;
-            }
-        }
-
-        // Right Turn (around x=ls, y=2R)
-        let dx_r = x - ls;
-        let dy_r = y_in_cycle - r * from_f64::<T>(2.0);
-        let d_sq_r = dx_r * dx_r + dy_r * dy_r;
-        if x >= ls && d_sq_r >= inner_r_sq && d_sq_r <= outer_r_sq {
-            return true;
-        }
-
-        // Left Turn (around x=0, y=4R/0)
-        // Check both top of this cycle and bottom of next
-        let dy_l_top = y_in_cycle - four_r;
-        let d_sq_l_top = x * x + dy_l_top * dy_l_top;
-        if x <= scalar::zero() && d_sq_l_top >= inner_r_sq && d_sq_l_top <= outer_r_sq {
-            return true;
-        }
-
-        let d_sq_l_bot = x * x + y_in_cycle * y_in_cycle;
-        if x <= scalar::zero() && d_sq_l_bot >= inner_r_sq && d_sq_l_bot <= outer_r_sq {
-            return true;
-        }
-
-        false
-    }
-}
-
-// ============================================================================
-// Mixing Models
-// ============================================================================
-
-/// Analytical model for advection-diffusion mixing in straight sections
-///
-/// In a straight microchannel with two inlet streams, concentration
-/// evolves according to Fick's law:
-///
-/// ```text
-/// ∂c/∂t + u·∂c/∂x = D·∂²c/∂y²
-/// ```
-///
-/// The segregation variance for an inlet step concentration is:
-/// ```text
-/// σ²(t)/σ²(0) = (8/π²) Σ_{m=0}∞ exp[-2(2m+1)²π²Dt/w²] / (2m+1)²
-/// ```
-///
-/// where `w` is channel width and `D` is molecular diffusivity.  A mixing
-/// fraction `M = 1 - σ(t)/σ(0)` reaches 90% when the variance ratio is `0.01`.
-///
-/// # Theorem — Transverse Diffusion Mixing Fraction
-///
-/// For a two-stream step concentration in a straight channel with no-flux
-/// sidewalls, the Neumann eigenfunctions `cos(nπy/w)` diagonalize the
-/// transverse diffusion operator. Only odd modes are present, and each mode
-/// decays as `exp(-n²π²Dt/w²)`.  The normalized variance therefore equals the
-/// odd-mode series above, and `M(x) = 1 - sqrt(σ²(x/u)/σ²(0))`.
-///
-/// **Proof.** Expanding the centered inlet step `c(y,0)-c_mean` in the Neumann
-/// basis gives coefficients `a_n = -2 sin(nπ/2)/(nπ)`. Orthogonality gives
-/// `σ²(t)=Σ a_n² exp(-2n²π²Dt/w²)/2`, while `σ²(0)=1/4`. Substitution leaves
-/// the odd-mode series and the stated mixing fraction. ∎
-pub struct AdvectionDiffusionMixing<T: Cfd2dScalar + Copy> {
-    /// Channel width \[m]
-    pub width: T,
-    /// Mean flow velocity \[m/s]
-    pub velocity: T,
-    /// Diffusion coefficient [m²/s]
-    pub diffusion_coeff: T,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> AdvectionDiffusionMixing<T> {
-    /// Create mixing model
-    pub fn new(width: T, velocity: T, diffusion_coeff: T) -> Self {
-        Self {
-            width,
-            velocity,
-            diffusion_coeff,
-        }
-    }
-
-    /// Calculate Peclet number
-    ///
-    /// Pe = u·w / D
-    pub fn peclet_number(&self) -> T {
-        (self.velocity * self.width) / (self.diffusion_coeff + from_f64::<T>(1e-15))
-    }
-
-    #[inline]
-    fn variance_ratio_from_fourier_number(fourier: T) -> T {
-        if fourier <= scalar::zero() {
-            return scalar::one();
-        }
-
-        let pi = from_f64::<T>(PI);
-        let pi_sq = pi * pi;
-        let eight_over_pi_sq = from_f64::<T>(8.0) / pi_sq;
-        let two = from_f64::<T>(2.0);
-        let mut sum: T = scalar::zero();
-
-        for mode in 0..256 {
-            let n = scalar::from_usize::<T>(2 * mode + 1);
-            let n_sq = n * n;
-            sum += FloatElement::exp(-two * n_sq * pi_sq * fourier) / n_sq;
-        }
-
-        eight_over_pi_sq * sum
-    }
-
-    /// Calculate advective length required for 90% variance-based mixing.
-    ///
-    /// The method solves `σ²(t)/σ²(0)=0.01` for `Fo = Dt/w²` by bisection of
-    /// the closed-form eigenfunction series and returns `L90 = u w² Fo90 / D`.
-    pub fn mixing_length_90_percent(&self) -> T {
-        if self.width <= scalar::zero()
-            || self.velocity <= scalar::zero()
-            || self.diffusion_coeff <= scalar::zero()
-        {
-            return scalar::zero();
-        }
-
-        let target_variance_ratio = from_f64::<T>(0.01);
-        let mut lo: T = scalar::zero();
-        let mut hi = from_f64::<T>(0.5);
-
-        for _ in 0..80 {
-            let mid = (lo + hi) / from_f64::<T>(2.0);
-            if Self::variance_ratio_from_fourier_number(mid) > target_variance_ratio {
-                lo = mid;
-            } else {
-                hi = mid;
-            }
-        }
-
-        let fo_90 = hi;
-        self.velocity * self.width * self.width * fo_90 / self.diffusion_coeff
-    }
-
-    /// Calculate mixing time to achieve 90% homogeneity
-    ///
-    /// t_mix = L_mix / u
-    pub fn mixing_time_90_percent(&self) -> T {
-        let l_mix = self.mixing_length_90_percent();
-        if self.velocity <= scalar::zero() {
-            scalar::zero()
-        } else {
-            l_mix / self.velocity
-        }
-    }
-
-    /// Estimate concentration profile at position x
-    ///
-    /// For two inlets with concentrations c_A and c_B, the concentration
-    /// at position x depends on diffusion progress:
-    ///
-    /// ```text
-    /// c(x, y) = c_A × erf(y / sqrt(4Dx/u)) + c_B × (1 - erf(...))
-    /// ```
-    ///
-    /// At the centerline (y=0), maximum mixing occurs there.
-    /// Returns fraction mixed at position x relative to channel width.
-    pub fn mixing_fraction(&self, x: T) -> T {
-        if x <= scalar::zero()
-            || self.width <= scalar::zero()
-            || self.velocity <= scalar::zero()
-            || self.diffusion_coeff <= scalar::zero()
-        {
-            return scalar::zero();
-        }
-
-        let fourier = self.diffusion_coeff * x / (self.velocity * self.width * self.width);
-        let variance_ratio = Self::variance_ratio_from_fourier_number(fourier);
-        scalar::one::<T>() - <T as NumericElement>::sqrt(variance_ratio)
-    }
-}
-
-// ============================================================================
-// Serpentine Mixing Solution
-// ============================================================================
-
-/// Solution for serpentine channel mixing
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct SerpentineMixingSolution<T: Cfd2dScalar + Copy> {
-    /// Inlet concentration (first inlet) [mol/m³]
-    pub c_inlet_a: T,
-    /// Inlet concentration (second inlet) [mol/m³]
-    pub c_inlet_b: T,
-    /// Peclet number (dimensionless)
-    pub peclet: T,
-    /// Mixing length for 90% homogeneity \[m]
-    pub l_mix_90: T,
-    /// Mixing time to 90% \[s]
-    pub t_mix_90: T,
-    /// Mixing fraction at outlet [0-1]
-    pub mixing_fraction_outlet: T,
-    /// Estimated viscous pressure drop \[Pa]
-    pub pressure_drop: T,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> SerpentineMixingSolution<T> {
-    /// Create solution from parameters
-    pub fn new(
-        geometry: &SerpentineGeometry<T>,
-        velocity: T,
-        diffusion_coeff: T,
-        c_inlet_a: T,
-        c_inlet_b: T,
-        viscosity: T,
-        density: T,
-    ) -> Self {
-        let mixing_model = AdvectionDiffusionMixing::new(geometry.width, velocity, diffusion_coeff);
-        let pe = mixing_model.peclet_number();
-        let l_mix = mixing_model.mixing_length_90_percent();
-        let t_mix = mixing_model.mixing_time_90_percent();
-
-        let total_length = geometry.total_length();
-        let mixing_frac = mixing_model.mixing_fraction(total_length);
-
-        // Pressure drop: ΔP = f × (L/D_h) × (1/2)ρu²
-        // For laminar flow, friction factor f = 64/Re
-        let d_h = from_f64::<T>(2.0) * (geometry.width * geometry.height)
-            / (geometry.width + geometry.height);
-        let re = (density * velocity * d_h) / viscosity;
-        let f = from_f64::<T>(64.0) / re.max(from_f64::<T>(1.0));
-        let dynamic_pressure = from_f64::<T>(0.5) * density * velocity * velocity;
-        let dp = f * (total_length / d_h) * dynamic_pressure;
-
-        Self {
-            c_inlet_a,
-            c_inlet_b,
-            peclet: pe,
-            l_mix_90: l_mix,
-            t_mix_90: t_mix,
-            mixing_fraction_outlet: mixing_frac,
-            pressure_drop: dp,
-        }
-    }
-
-    /// Check if mixing is achieved at outlet
-    ///
-    /// Typically consider "mixed" if fraction > 0.9 (90%)
-    pub fn is_well_mixed(&self) -> bool {
-        self.mixing_fraction_outlet > from_f64::<T>(0.9)
-    }
-
-    /// Estimate outlet concentration (assuming complete mixing)
-    ///
-    /// For equal volume flows: c_outlet = (c_A + c_B) / 2
-    pub fn estimated_outlet_concentration(&self) -> T {
-        (self.c_inlet_a + self.c_inlet_b) / from_f64::<T>(2.0)
-    }
-}
-
-// ============================================================================
-// Serpentine Validator
-// ============================================================================
-
-/// Validator for serpentine mixing channels
-pub struct SerpentineValidator<T: Cfd2dScalar + Copy> {
-    geometry: SerpentineGeometry<T>,
-}
-
-impl<T: Cfd2dScalar + Copy + FloatElement> SerpentineValidator<T> {
-    /// Create new validator
-    pub fn new(geometry: SerpentineGeometry<T>) -> Self {
-        Self { geometry }
-    }
-
-    /// Validate mixing efficiency against expected values
-    ///
-    /// # Validation Criteria
-    ///
-    /// For well-designed serpentines:
-    /// - Mixing length < total channel length
-    /// - Mixing fraction at outlet > 90%
-    /// - Pressure drop < 100 kPa (for ~mm-scale channels)
-    pub fn validate_mixing(
-        &self,
-        solution: &SerpentineMixingSolution<T>,
-    ) -> Result<SerpentineValidationResult<T>, String> {
-        let total_length = self.geometry.total_length();
-        let l_mix = solution.l_mix_90;
-
-        // Check if mixing length is achieved within channel
-        let achievable = l_mix < total_length * from_f64::<T>(2.0);
-
-        // Check mixing fraction at outlet
-        let well_mixed = solution.is_well_mixed();
-
-        // Check pressure drop is reasonable
-        let dp_reasonable = solution.pressure_drop < from_f64::<T>(1e5); // < 100 kPa
-
-        let validation_passed = achievable && well_mixed && dp_reasonable;
-
-        let mut error_message = None;
-        if !achievable {
-            error_message = Some("Mixing length exceeds 2× total channel length".to_string());
-        } else if !well_mixed {
-            error_message = Some(format!(
-                "Mixing fraction {:.2}% < 90% at outlet",
-                eunomia::NumericElement::to_f64(
-                    solution.mixing_fraction_outlet * from_f64::<T>(100.0),
-                )
-            ));
-        } else if !dp_reasonable {
-            error_message = Some(format!(
-                "Pressure drop {:.0} Pa > 100 kPa",
-                eunomia::NumericElement::to_f64(solution.pressure_drop)
-            ));
-        }
-
-        Ok(SerpentineValidationResult {
-            validation_passed,
-            mixing_fraction_outlet: solution.mixing_fraction_outlet,
-            error_message,
-        })
-    }
-}
-
-/// Validation result for serpentine
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SerpentineValidationResult<T: Cfd2dScalar + Copy> {
-    /// Validation passed
-    pub validation_passed: bool,
-    /// Measured mixing fraction at outlet
-    pub mixing_fraction_outlet: T,
-    /// Error message if validation failed
-    pub error_message: Option<String>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::AdvectionDiffusionMixing;
-    use eunomia::NumericElement;
-
-    #[test]
-    fn transverse_diffusion_series_starts_unmixed() {
-        let model = AdvectionDiffusionMixing::new(200e-6_f64, 0.01, 1.0e-10);
-        assert_eq!(model.mixing_fraction(0.0), 0.0);
-    }
-
-    #[test]
-    fn mixing_length_reaches_ninety_percent_by_definition() {
-        let model = AdvectionDiffusionMixing::new(200e-6_f64, 0.01, 1.0e-10);
-        let l90 = model.mixing_length_90_percent();
-        let fraction = model.mixing_fraction(l90);
-
-        assert!(
-            <f64 as NumericElement>::abs(fraction - 0.9) < 5.0e-12,
-            "Expected eigenfunction L90 to produce 90% mixing, got {fraction:.15}"
-        );
-    }
-
-    #[test]
-    fn mixing_length_scales_with_velocity_and_inverse_diffusivity() {
-        let baseline = AdvectionDiffusionMixing::new(200e-6_f64, 0.01, 1.0e-10);
-        let faster = AdvectionDiffusionMixing::new(200e-6_f64, 0.02, 1.0e-10);
-        let more_diffusive = AdvectionDiffusionMixing::new(200e-6_f64, 0.01, 2.0e-10);
-
-        assert!(
-            <f64 as NumericElement>::abs(
-                faster.mixing_length_90_percent() / baseline.mixing_length_90_percent() - 2.0,
-            ) < 1.0e-12
-        );
-        assert!(
-            <f64 as NumericElement>::abs(
-                more_diffusive.mixing_length_90_percent() / baseline.mixing_length_90_percent()
-                    - 0.5,
-            ) < 1.0e-12
-        );
-    }
-}
+pub use validation::{SerpentineValidationResult, SerpentineValidator};
